break up the PipelineDecorator into smaller purpose-driven classes
