+++ src/main/groovy/org/boozallen/plugins/jte/init/JteBlockWrapper.groovy
@@ -1,0 +1,42 @@
+/*
+    Copyright 2018 Booz Allen Hamilton
+
+    Licensed under the Apache License, Version 2.0 (the "License");
+    you may not use this file except in compliance with the License.
+    You may obtain a copy of the License at
+
+        http://www.apache.org/licenses/LICENSE-2.0
+
+    Unless required by applicable law or agreed to in writing, software
+    distributed under the License is distributed on an "AS IS" BASIS,
+    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+    See the License for the specific language governing permissions and
+    limitations under the License.
+*/
+package org.boozallen.plugins.jte.init
+
+/**
+ * Represents the jte{} block in the Pipeline Configuration
+ */
+@SuppressWarnings("PropertyName")
+class JteBlockWrapper {
+
+    String pipeline_template = null
+    Boolean allow_scm_jenkinsfile = true
+    Boolean permissive_initialization = false
+    Boolean reverse_library_resolution = false
+
+    static LinkedHashMap getSchema(){
+        return [
+            fields: [
+                optional: [
+                    allow_scm_jenkinsfile: Boolean,
+                    pipeline_template: String,
+                    permissive_initialization: Boolean,
+                    reverse_library_resolution: Boolean
+                ]
+            ]
+        ]
+    }
+
+}
