+++ src/main/groovy/org/boozallen/plugins/jte/init/governance/config/dsl/PipelineConfigurationObject.groovy
@@ -15,7 +15,8 @@
 */
 package org.boozallen.plugins.jte.init.governance.config.dsl
 
-import org.boozallen.plugins.jte.init.PipelineDecorator
+import org.boozallen.plugins.jte.init.JteBlockWrapper
+
 import org.boozallen.plugins.jte.util.TemplateLogger
 import org.codehaus.groovy.runtime.InvokerHelper
 import org.jenkinsci.plugins.workflow.flow.FlowExecutionOwner
@@ -39,8 +40,8 @@ class PipelineConfigurationObject implements Serializable{
         this.flowOwner = flowOwner
     }
 
-    PipelineDecorator.JteBlockWrapper getJteBlockWrapper(){
-        return (config.jte ?: [:]) as PipelineDecorator.JteBlockWrapper
+    JteBlockWrapper getJteBlockWrapper(){
+        return (config.jte ?: [:]) as JteBlockWrapper
     }
 
     PipelineConfigurationObject plus(PipelineConfigurationObject child){
