+++ src/main/groovy/org/boozallen/plugins/jte/job/TemplateFlowDefinition.groovy
@@ -17,11 +17,14 @@ package org.boozallen.plugins.jte.job
 
 import static org.jenkinsci.plugins.workflow.cps.persistence.PersistenceContext.JOB
 
+import org.boozallen.plugins.jte.init.PipelineConfigurationAggregator
+import org.boozallen.plugins.jte.init.PipelineTemplateResolver
+import org.boozallen.plugins.jte.init.governance.config.dsl.PipelineConfigurationObject
+import org.boozallen.plugins.jte.init.primitives.TemplatePrimitiveInjector
 import hudson.model.Action
 import hudson.model.Item
 import hudson.model.Queue
 import hudson.model.TaskListener
-import org.boozallen.plugins.jte.init.PipelineDecorator
 import org.jenkinsci.plugins.workflow.cps.CpsFlowExecution
 import org.jenkinsci.plugins.workflow.cps.persistence.PersistIn
 import org.jenkinsci.plugins.workflow.flow.FlowDefinition
@@ -34,8 +37,8 @@ import org.jenkinsci.plugins.workflow.job.WorkflowRun
 import org.jenkinsci.plugins.workflow.cps.CpsFlowFactoryAction2
 
 /**
- * A custom {@link FlowDefinition} that initializes the pipeline using the {@link PipelineDecorator} prior to returning
- * a {@link CpsFlowExecution} representing the pipeline template
+ * Defines a JTE Pipeline.
+ * Performs JTE's Pipeline Initialization Process before creating the CpsFlowExecution for the WorkflowRun
  */
 @PersistIn(JOB)
 abstract class TemplateFlowDefinition extends FlowDefinition {
@@ -51,21 +54,31 @@ abstract class TemplateFlowDefinition extends FlowDefinition {
 
     @Override
     FlowExecution create(FlowExecutionOwner owner, TaskListener listener, List<? extends Action> actions) throws Exception {
+        // if this is a replay run, just skip initialization
         for (Action a : actions) {
             if (a instanceof CpsFlowFactoryAction2) {
                 return ((CpsFlowFactoryAction2) a).create(this, owner, actions)
             }
         }
+        String template = initializePipeline(owner)
         FlowDurabilityHint hint = determineFlowDurabilityHint(owner)
-        String template = initializeJTE(owner)
-
         return new CpsFlowExecution(template, true, owner, hint)
     }
 
-    private String initializeJTE(FlowExecutionOwner owner){
-        PipelineDecorator decorator = new PipelineDecorator(owner)
-        decorator.initialize() // runs the initialization process for JTE
-        String template = decorator.getTemplate()
+    /**
+     * Performs JTE's Pipeline Initialization Process
+     * @param owner
+     * @return the Pipeline Template for the Run
+     */
+    String initializePipeline(FlowExecutionOwner owner){
+        // Step 1: Aggregate pipeline configs
+        PipelineConfigurationAggregator configurationAggregator = new PipelineConfigurationAggregator(owner)
+        PipelineConfigurationObject config = configurationAggregator.aggregate()
+        // Step 2: Invoke TemplatePrimitiveInjectors
+        TemplatePrimitiveInjector.orchestrate(owner, config)
+        // Step 3: determine the Pipeline Template
+        PipelineTemplateResolver templateResolver = new PipelineTemplateResolver(owner)
+        String template = templateResolver.resolve(config)
         return template
     }
